Cloth and User Service implementation

+++ src/main/java/com/example/SmartCloset/service/ClothService.java
@@ -1,15 +1,13 @@
 package com.example.SmartCloset.service;
 
 import com.example.SmartCloset.model.ClosetEnum.ClothesColor;
-import com.example.SmartCloset.model.ClosetEnum.TPO;
 import com.example.SmartCloset.model.Cloth;
-
 import java.util.ArrayList;
 import java.util.HashMap;
 
 public interface ClothService {
 
-    HashMap<ClothesColor, Float> getColorDistribution(ArrayList<Cloth> likedClothes);
+    HashMap<ClothesColor, Integer> getColorDistribution(ArrayList<Cloth> likedClothes);
 
     // TODO: 2022/07/15 Category2 에 대한 distribution
